Move tests of consensus to its own module
